pyproject.toml
[project]
name = "pdfaker"
version = "0.1.0"
description = "App to combine JPG, PDF files into one"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "reportlab>=4.3.1",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
pdfaker = "main:main"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }